+++ src/osx/shake.c
@@ -8,7 +8,7 @@
 
 #include "shake.h"
 #include "./shake_private.h"
-#include "../common/helpers.h"
+#include "../common/error.h"
 
 listElement *listHead;
 unsigned int numOfDevices;
